requirements.txt
discord.py==2.6.4
Flask==3.1.2
protobuf==6.33.1
psutil==7.1.3
py_dactyl==2.0.4
python-dotenv==1.2.1
pytz==2025.2
Requests==2.32.5
google-genai
